fix: change renderChatMessage hook to obfuscate hidden rolls properly
Currently private dice rolls show the effect to everyone instead of obfuscating it. This change
hides the effect and does not colorize the result if the user is not allowed to view it. There was
also a bug ignoring the user's choice of roll mode in the roll dialog, which this commit fixes.

Fix #321