add: Add the ability to display Lifeblood and Stamina
This change is for CD compatibility.  New characteristics are added: stamina and lifeblood. Currently the values are just placeholders and do no interact with damage.  As part of this mod, two new sub-values of "inputCurrent" have been created to allow the user to enter the current value for these characteristics (normally this is a calculated number from value and damage).  In the future it should be updated to auto calc damage.