feat: add weapons handling modifiers (#1236)
This update introduces weapons handling modifiers.  These modifiers are applied as bonuses or penalties depending on a specific actor characteristic.  This type of modifier is used in Classic Traveller (and possibly other rulesets).  A new string field has been added to the weapons sheet.  An example modifier looks like "DEX 6/-2 10/+1".   It translates as - if the actor has a dexterity of 6 or less, a -2 DM is applied to the weapon roll.  If the actor has a dexterity of 10 or more, a +1 DM is applied.  From 7 and 9, no DM is applied.  Note that the abbreviated characteristic name must match the actor's sheet.  So if you have custom names, you need to use those custom names.  Thanks to xdy and IanE for suggesting.  It has been two+ years, but I've finally gotten to this issue (the oldest one on GitHub).