feat: Add chat context items for damage and healing menu (#1363)
This release adds healing and damage actions to the chat card context menu. Similar to the way it works in the dnd5e system. To use these new actions, select tokens that you want the damage or healing applied (select not target).  Right-click any roll chat message to see the actions options. Click the option desired and all selected tokens will be affected.  There are more details of this feature (and also other methods of applying healing and damage) on a new page of the Wiki.

This release also includes more Spanish localization updates from @ForjaSalvaje